+++ packages/storage/src/providers/s3/utils/constants.ts
@@ -25,3 +25,5 @@ export const STORAGE_INPUT_KEY = 'key';
 export const STORAGE_INPUT_PATH = 'path';
 
 export const DEFAULT_DELIMITER = '/';
+
+export const CHECKSUM_ALGORITHM_CRC32 = 'crc-32';
