+++ packages/storage/src/providers/s3/utils/getCombinedCrc32.native.ts
@@ -6,6 +6,18 @@ import { getDataChunker } from '../apis/uploadData/multipart/getDataChunker';
 
 import { calculateContentCRC32 } from './crc32';
 
+/**
+ * Calculates a combined CRC32 checksum for the given data.
+ *
+ * This function chunks the input data, calculates CRC32 for each chunk,
+ * and then combines these checksums into a single value.
+ *
+ * @async
+ * @param {StorageUploadDataPayload} data - The data to calculate the checksum for.
+ * @param {number | undefined} size - The size of each chunk. If undefined, a default chunk size will be used.
+ * @returns {Promise<string>} A promise that resolves to a string containing the combined CRC32 checksum
+ *                            and the number of chunks, separated by a hyphen.
+ */
 export const getCombinedCrc32 = async (
 	data: StorageUploadDataPayload,
 	size: number | undefined,
